+++ spotify/__init__.py
@@ -122,6 +122,7 @@ from spotify.offline import *  # noqa
 from spotify.playlist import *  # noqa
 from spotify.search import *  # noqa
 from spotify.session import *  # noqa
+from spotify.sink import *  # noqa
 from spotify.social import *  # noqa
 from spotify.toplist import *  # noqa
 from spotify.track import *  # noqa